Backend/requirements.txt
fastapi
uvicorn
python-multipart
jinja2
gitpython
ruff
pytest